Finished polish
Implemented polishing elements. Added a timer that tracks across scenes. Put in a start and end screen that teaches player what to do and allows them to close the game. Put in a respawn button. Adjusted the dash so it is easier to control.  Added a death plane to each level to ensure player doesn't fall forever.  Adjusted background functions to ensure game runs smoothly like scene changing and some physics elements with the dash. Hopefully this is the final commit.